NO-ISSUE: more I18 N
- bug fixes: crash on redo and wrong replace fixed.

+++ FindAndReplace.c
@@ -668,7 +668,7 @@ int EdReplaceText(int scope, int action, int flags)
 	MARK		*markstart,*Markend;
 	register MARK *markend;
 	LINE		*oldxpnd = 0;
-	size_t newlen;
+	long newlen;
 	register	maxlen,delta;
 	RE_MATCH	match;
 	WINFO* wp = ww_getCurrentEditorWindow();
@@ -686,7 +686,7 @@ int EdReplaceText(int scope, int action, int flags)
 		regex_compileWithDefault(_currentSearchAndReplaceParams.searchPattern);
 		find_initializeReplaceByExpression(_currentSearchAndReplaceParams.replaceWith);
 	}
-	newlen = _currentReplacementPattern.preparedReplacementString ? strlen(_currentReplacementPattern.preparedReplacementString) : 0;
+	newlen = _currentReplacementPattern.preparedReplacementString ? (long)strlen(_currentReplacementPattern.preparedReplacementString) : 0;
 
 	/* call before assigning firstline	*/
 	hist_saveString(SEARCH_AND_REPLACE, _currentSearchAndReplaceParams.replaceWith);
